Clean up some compiler warnings, add method for settings service in service accessor

+++ src/app/activity/gui/game_activity_item.cpp
@@ -19,7 +19,7 @@ void GameActivityItem::setContentHash(const QString &contentHash) {
   m_timesPlayed = 0;
   m_totalSecondsPlayed = 0;
   for (const auto &session : playSessions) {
-    printf("Play session: %s, %lu, %lu, %lu\n", session.contentHash.c_str(),
+    printf("Play session: %s, %llu, %llu, %llu\n", session.contentHash.c_str(),
            session.startTime, session.endTime, session.unpausedDurationMillis);
     m_timesPlayed++;
     m_totalSecondsPlayed += session.unpausedDurationMillis / 1000;
@@ -31,7 +31,7 @@ void GameActivityItem::setContentHash(const QString &contentHash) {
   emit contentHashChanged();
 }
 
-int GameActivityItem::getTotalSecondsPlayed() const {
+uint64_t GameActivityItem::getTotalSecondsPlayed() const {
   return m_totalSecondsPlayed;
 }
